// apps/web/src/lib/api/generated/main/model/user-npc-kills-response-dto-output-npcs-item.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type UserNpcKillsResponseDtoOutputNpcsItem = {
  npcId: number;
  npcName: string;
  npcType: string;
  npcLvl: number;
  /** @nullable */
  npcProf: string | null;
  /** @nullable */
  npcIcon: string | null;
  totalKills: number;
};